+++ Editor/Attribute/InspectorInjection/IInspectorInjectionBackend.cs
@@ -1,0 +1,13 @@
+#if UNITY_EDITOR
+using System;
+
+namespace Jeomseon.Attribute.Editor
+{
+    internal interface IInspectorInjectionBackend : IDisposable
+    {
+        string Name { get; }
+        bool IsSupported { get; }
+        void Start();
+    }
+}
+#endif